--- src/rust/klep-mutation-v2/Cargo.toml
@@ -1,50 +1,0 @@
-[package]
-name = "klep-mutation-v2"
-version = "0.2.0"
-edition = "2021"
-
-[[bin]]
-name = "klep-mutation-v2"
-path = "src/main.rs"
-
-[dependencies]
-# CLI and configuration
-clap = { version = "4.0", features = ["derive"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-
-# Async runtime
-tokio = { version = "1.0", features = ["full"] }
-futures = "0.3"
-
-# Simple regex-based parsing (more reliable than tree-sitter native compilation)
-regex = "1.10"
-
-# File operations
-walkdir = "2.0"
-tempfile = "3.8"
-
-# Parallelization 
-rayon = "1.8"
-crossbeam = "0.8"
-dashmap = "6.0"
-
-# Error handling
-anyhow = "1.0"
-thiserror = "1.0"
-
-# Performance and utilities
-once_cell = "1.19"
-uuid = { version = "1.6", features = ["v4"] }
-
-# Date/time for reports
-chrono = { version = "0.4", features = ["serde"] }
-
-# Performance optimization dependencies
-memmap2 = "0.9"        # Memory-mapped file I/O
-ahash = "0.8"          # Fast non-crypto hashing
-lru = "0.12"           # LRU cache for compiled artifacts
-
-[dev-dependencies]
-tempfile = "3.8"
-sha256 = "1.4"
